Make navigation and search component responsive
