+++ src/App.js
@@ -1,6 +1,7 @@
 import { Route, Routes } from "react-router-dom";
 import "./App.css";
-import Navigation from "./pages/Navigation";
+import Navigation from "./components/Navigation";
+ 
 
 function App() {
   return (
