--- src/pages/Navigation.js
@@ -1,13 +1,0 @@
-import DesktopView from "../components/DesktopView";
-import MobileView from "../components/MobileView";
-
-const Navigation = () => {
-  return (
-    <div>
-        <MobileView />
-        <DesktopView />
-    </div>
-  );
-};
-
-export default Navigation;
